+++ pyproject.toml
@@ -17,6 +17,7 @@ dxgb-ext-bench = "dxgb_bench.dxgb_ext_bench:cli_main"
 dxgb-to-csv = "dxgb_bench.dxgb_to_csv:cli_main"
 dxgb-bench = "dxgb_bench.dxgb_bench:cli_main"
 dxgb-dist-bench = "dxgb_bench.dxgb_dist_bench:cli_main"
+_dxgb-dist-impl = "dxgb_bench.dxgb_dist_impl:cli_main"
 
 [tool.setuptools]
 include-package-data = true
